Issue #41: New Channel Owner

+++ groupease-server/src/main/java/io/github/groupease/channel/DefaultChannelService.java
@@ -10,6 +10,10 @@ import javax.inject.Inject;
 import com.codahale.metrics.annotation.Timed;
 import com.google.common.base.Strings;
 import com.google.inject.persist.Transactional;
+import io.github.groupease.db.MemberDao;
+import io.github.groupease.model.Member;
+import io.github.groupease.user.GroupeaseUser;
+import io.github.groupease.user.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -24,6 +28,8 @@ public class DefaultChannelService implements ChannelService {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
     private final ChannelDao channelDao;
+    private final MemberDao memberDao;
+    private final UserService userService;
 
     /**
      * Injectable constructor.
@@ -32,9 +38,13 @@ public class DefaultChannelService implements ChannelService {
      */
     @Inject
     public DefaultChannelService(
-            @Nonnull ChannelDao channelDao
+            @Nonnull ChannelDao channelDao,
+            @Nonnull MemberDao memberDao,
+            @Nonnull UserService userService
     ) {
         this.channelDao = requireNonNull(channelDao);
+        this.memberDao = requireNonNull(memberDao);
+        this.userService = requireNonNull(userService);
     }
 
     @Nonnull
@@ -75,13 +85,26 @@ public class DefaultChannelService implements ChannelService {
         LOGGER.debug("DefaultChannelService.update({}) called.", toUpdate);
 
         requireNonNull(toUpdate);
+        requireNonNull(toUpdate.getId());
 
         if (Strings.isNullOrEmpty(toUpdate.getName())) {
             throw new ChannelNameMissingException("Channel name is required.");
         }
 
-        /* Confirm toUpdate exists and current user has access. */
-        channelDao.getById(toUpdate.getId());
+        /* Confirm current user is owner of channel. */
+
+        /* Get current user. */
+        GroupeaseUser currentUser = getCurrentUser();
+
+        /* Get channel member. */
+        Member member = memberDao.getById(
+                currentUser.getId(),
+                toUpdate.getId()
+        );
+
+        if (member == null || !member.isOwner()) {
+            throw new ChannelEditByNonOwnerException();
+        }
 
         return channelDao.update(toUpdate);
     }
@@ -105,7 +128,21 @@ public class DefaultChannelService implements ChannelService {
             throw new ChannelNameMissingException("Channel name is required.");
         }
 
-        return channelDao.create(toCreate);
+        /* Create channel. */
+        Channel newChannel = channelDao.create(toCreate);
+
+        /* Get current user. */
+        GroupeaseUser currentUser = getCurrentUser();
+
+        /* Add current user as channel member and owner. */
+        memberDao.create(
+                currentUser.getId(),
+                newChannel.getId(),
+                true
+        );
+
+        /* Return the new channel. */
+        return newChannel;
     }
 
     @Override
@@ -116,10 +153,28 @@ public class DefaultChannelService implements ChannelService {
     ) {
         LOGGER.debug("DefaultChannelService.delete({}) called.", id);
 
-        /* Confirm channel to delete exists and current user has access. */
-        channelDao.getById(id);
+        /* Confirm current user is owner of channel. */
 
+        /* Get current user. */
+        GroupeaseUser currentUser = getCurrentUser();
+
+        /* Get channel member. */
+        Member member = memberDao.getById(
+                currentUser.getId(),
+                id
+        );
+
+        if (member == null || !member.isOwner()) {
+            throw new ChannelEditByNonOwnerException();
+        }
+
+        /* Delete channel. */
         channelDao.delete(id);
+    }
+
+    private GroupeaseUser getCurrentUser() {
+        /* Get current user, ensuring saved in DB and refreshing from source. */
+        return userService.updateCurrentUser();
     }
 
 }
